Refactor XML handling with XElementBuilder class
This commit introduces the `XElementBuilder` class to replace the previous `XElementNode` and `XElement` handling in the `DataDictionary` namespace.
